+++ CHANGELOG.md
@@ -8,6 +8,41 @@ Maintainer: Shiro Takeda
 Change logs of GAMS mode.
 ====================================================
 
+Version 6.16
+----------------------------------
+
+* The font-locking and autocompletion features have been improved. 
+
+* `company-mode`, which is a text completion framework for Emacs, has been
+  available in GAMS mode since version 6.13. If you want to use
+  company-mode in GAMS mode, 1) install company-mode by yourself, and 2)
+  add the following code to init.el
+
+```emacs-lisp
+(add-hook
+  'gams-mode-hook
+  #'(lambda () (company-mode))
+  )
+```
+
+Version 6.15
+----------------------------------
+
+* In GAMS-SIL mode, $label command is now treated in the same way as
+  $title command.
+  
+
+
+Version 6.14
+----------------------------------
+
+* Improved code for font locking. For this purpose, I have added the
+  `gams-commands-dollar.txt` file, which contains a list of dollar control
+  options. In addition, I changed the `gams-commands.txt` file to include
+  only commands other than dollar control options.
+
+
+
 Version 6.13
 ----------------------------------